chore : 하나의 클래스로 통일
